Refactor game registration and initialization logic
- moved game migrations and db upgrading to a different script in init_db.py instead of keeping it in the app creation and making loading heavy
- Otherwise gunicorn workers get timeout because of the time it takes to add games to db.
