+++ backend/run.py
@@ -4,7 +4,7 @@ import os
 app = create_app()
 
 
-initialize_app(app)
-
 if __name__ == "__main__":
+    if os.environ.get("WERKZEUG_RUN_MAIN") == "true":
+        initialize_app(app)
     app.run(host="0.0.0.0", port=5000, debug=True)
